Move concatenation tests out into their own suite.
