+++ MadnessTests/ParserTests.swift
@@ -61,25 +61,6 @@ final class ParserTests: XCTestCase {
 
 	// MARK: - Nonterminals
 
-	// MARK: Concatenation
-
-	let concatenation = %"x" ++ %"y"
-
-	func testConcatenationRejectsPartialParses() {
-		assertNil(concatenation("x"))
-	}
-
-	func testConcatenationParsesBothOperands() {
-		assertEqual(concatenation("xyz")?.1, "z")
-	}
-
-	func testConcatenationProducesPairsOfTerms() {
-		let parsed = concatenation("xy")
-		assertEqual(parsed?.0.0, "x")
-		assertEqual(parsed?.0.1, "y")
-	}
-
-
 	// MARK: Alternation
 
 	let alternation = %"x" | (%"y" --> const(1))
